fix value error when parsing iozone output.
Use split() method is more robust than counting whitespaces and extract digits manually.
The split() method splits a string using the input parameter as the delimiter string.
If no separator is specified (default separator is a whitespace) it works differently,
runs of consecutive whitespaces are regarded as a single separator.
So the result contains no empty strings.